requirements.txt
torch~=1.6.0
PyYAML~=5.3.1
torchvision~=0.7.0
Pillow~=8.0.1
torchsummary~=1.5.1
runx~=0.0.10